requirements.txt
# Core dependencies
python-dotenv>=1.0.0
openai>=1.0.0
mcp>=0.1.0
torch>=2.0.0
transformers>=4.40.0
peft>=0.10.0
accelerate>=0.27.0
bitsandbytes>=0.41.0  # Optional, for quantization
deepspeed>=0.14.0  # Optional, for multi-GPU
tqdm>=4.65.0
pyyaml>=6.0

# WandB and Weave for evaluation tracking
wandb>=0.15.0
weave>=0.50.0

# Async support
asyncio>=3.4.3

# Optional but recommended
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0